Adding sunset report figures and results for vanilla net

+++ patternRec/hw4/scripts/download_model_binary.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env sh
 
-python2 python/download_model_binary.py uris/${1}/
-mv uris/${1}/*.caffemodel snapshots/${1}_init.caffemodel
+if [ ! -f snapshots/${1}_init.caffemodel ]; then
+  python2 python/download_model_binary.py uris/${1}/
+  mv uris/${1}/*.caffemodel snapshots/${1}_init.caffemodel
+fi
